Fixed issue with # of threads.
I didn't do an extensive test 'couse it'll require a lot of time, but it
didn't crash while downloading some not "easy" artist, so I'm pretty
confident to say that it works.
It's still asking for more than 3GiB, but it doesn't seem like it needs
to go any higher.